test(voice): address the review findings on the contract journeys
Six review threads from #8056.

- Split the 850-line step file into six focused modules under the
  300-line limit for a new file, keeping steps.ts as a barrel. No
  assertion changed while moving.
- Fail immediately when provider setup returns a non-success status.
  Playwright does not throw on its own, so later steps could run
  without a confirmed provider.
- Stop falling back to the first project when the pinned slug is
  missing, which could write real provider credentials into an
  unrelated project. Throw and name the slugs that were available.
- Assert audio on every conversation part rather than on the first
  match, and assert the part count is above zero so the loop cannot
  pass on an empty list.
- Derive each test timeout from its own phase constants plus a named
  setup margin, instead of a hardcoded 300 seconds that was already
  below the waits inside.
- Give each run a unique agent name, so a stale card from an earlier
  run in the same project cannot be selected.
